+++ code/src/models/__init__.py
@@ -3,6 +3,6 @@ from .chroma_input_model import ChromaIntegration
 from .milvus_input_model import MilvusIntegration
 from .opensearch_input_model import OpensearchIntegration
 from .pgvector_input_model import PgvectorIntegration
-from .pinecone_input_model import EmbeddingsProvider, PineconeIntegration
+from .pinecone_input_model import PineconeIntegration
 from .qdrant_input_model import QdrantIntegration
 from .weaviate_input_model import WeaviateIntegration